Split graphqlfetch.go and run.go by concern, fix stale doc comments
graphqlfetch.go held three independent GraphQL fetch paths plus shared
query/response plumbing under one generic name; split into
openprsfetch.go, mergedprsfetch.go, getprsbyreffetch.go, graphqlquery.go
and an extended graphqlmodels.go, mirroring how the tests already
grouped by fetch path.

run.go mixed the reminder-message path with the PR tracker canvas
refresh path; canvas.go now holds the latter, matching canvas_test.go's
existing scenario grouping.

Also fixes four package doc comments that no longer described what
their packages do, including inputhelpers.go's leftover "Package
utilities".

No behavior change: pure file reorg plus comment fixes.
