--- internal/apiclients/githubclient/graphqlfetch.go
@@ -1,707 +1,0 @@
-package githubclient
-
-import (
-	"cmp"
-	"context"
-	"encoding/json"
-	"errors"
-	"fmt"
-	"log"
-	"slices"
-	"strings"
-	"time"
-
-	"github.com/hellej/pr-slack-reminder-action/internal/models"
-	"github.com/hellej/pr-slack-reminder-action/internal/utilities"
-	"golang.org/x/sync/errgroup"
-)
-
-const (
-	openPullRequestState   = "open"
-	closedPullRequestState = "closed"
-)
-
-// GraphQL states of a PR node. CLOSED and MERGED are both "closed" to the pipeline.
-const (
-	closedNodeState = "CLOSED"
-	mergedNodeState = "MERGED"
-)
-
-const notFoundErrorType = "NOT_FOUND"
-
-const rateLimitField = "rateLimit"
-
-// labels are capped at GitHub's maximum page size; REST returned them unpaginated.
-const openPullRequestsFragment = `fragment prs on Repository {
-  pullRequests(states: OPEN, first: 100, orderBy: {field: CREATED_AT, direction: DESC}) {
-    nodes {
-      number title url isDraft createdAt updatedAt headRefOid
-      author { login __typename ... on User { name } }
-      labels(first: 100) { nodes { name } }
-    }
-  }
-}`
-
-type graphqlQuery struct {
-	text              string
-	variables         map[string]any
-	aliases           []string
-	repositoryByAlias map[string]models.Repository
-	referenceByAlias  map[string]models.PullRequestRef // PR queries only
-}
-
-// Aliases cannot be variables, so they are generated into the query text; owner and name are
-// always passed as variables.
-func buildListOpenPRsQuery(repositories []models.Repository) graphqlQuery {
-	query := graphqlQuery{
-		variables:         map[string]any{},
-		aliases:           make([]string, len(repositories)),
-		repositoryByAlias: map[string]models.Repository{},
-	}
-	variableDeclarations := make([]string, 0, 2*len(repositories))
-	repositorySelections := make([]string, len(repositories))
-
-	for index, repository := range repositories {
-		alias := fmt.Sprintf("r%d", index)
-		ownerVariable := fmt.Sprintf("owner%d", index)
-		nameVariable := fmt.Sprintf("name%d", index)
-
-		query.aliases[index] = alias
-		query.repositoryByAlias[alias] = repository
-		query.variables[ownerVariable] = repository.Owner
-		query.variables[nameVariable] = repository.Name
-		variableDeclarations = append(
-			variableDeclarations, "$"+ownerVariable+":String!", "$"+nameVariable+":String!",
-		)
-		repositorySelections[index] = fmt.Sprintf(
-			"  %s: repository(owner:$%s,name:$%s){ ...prs }", alias, ownerVariable, nameVariable,
-		)
-	}
-
-	query.text = assembleQuery(variableDeclarations, repositorySelections, openPullRequestsFragment)
-	return query
-}
-
-// search returns issues and PRs, so the PR fields are selected through an inline fragment and
-// need no named one. The selection is narrower than the open-PR listing's: a merged row shows no
-// draft marker, no activity and no head commit.
-const mergedPullRequestSearchSelection = `{ issueCount nodes { ... on PullRequest {
-    number title url createdAt mergedAt
-    author { login __typename ... on User { name } }
-    labels(first: 100){ nodes { name } }
-  } } }`
-
-// GitHub's maximum page size for a search connection, and the point past which a repository's
-// merges no longer fit in one page.
-const mergedPRSearchPageSize = 100
-
-const mergedSinceDayLayout = "2006-01-02"
-
-// One aliased search per repository, so each result set maps back to its repository without
-// selecting the repository on every node.
-func buildMergedPRsSearchQuery(
-	repositories []models.Repository, mergedSince time.Time,
-) graphqlQuery {
-	query := graphqlQuery{
-		variables:         map[string]any{},
-		aliases:           make([]string, len(repositories)),
-		repositoryByAlias: map[string]models.Repository{},
-	}
-	variableDeclarations := make([]string, len(repositories))
-	searchSelections := make([]string, len(repositories))
-
-	for index, repository := range repositories {
-		alias := fmt.Sprintf("s%d", index)
-		searchQueryVariable := fmt.Sprintf("q%d", index)
-
-		query.aliases[index] = alias
-		query.repositoryByAlias[alias] = repository
-		query.variables[searchQueryVariable] = mergedPRSearchQueryString(repository, mergedSince)
-		variableDeclarations[index] = "$" + searchQueryVariable + ":String!"
-		searchSelections[index] = fmt.Sprintf(
-			"  %s: search(query:$%s, type: ISSUE, first: %d)%s",
-			alias, searchQueryVariable, mergedPRSearchPageSize, mergedPullRequestSearchSelection,
-		)
-	}
-
-	query.text = assembleQuery(variableDeclarations, searchSelections, "")
-	return query
-}
-
-// The cutoff is a day rather than a timestamp, so the search returns up to one extra day of
-// merges; the exact cut happens client-side.
-func mergedPRSearchQueryString(repository models.Repository, mergedSince time.Time) string {
-	return fmt.Sprintf(
-		"repo:%s is:pr is:merged merged:>=%s",
-		repository.GetPath(), mergedSince.UTC().Format(mergedSinceDayLayout),
-	)
-}
-
-type searchResultsNode struct {
-	IssueCount int               `json:"issueCount"`
-	Nodes      []pullRequestNode `json:"nodes"`
-}
-
-// Any error fails the whole merged fetch: search reports a nonexistent repository, one the token
-// cannot read and an empty window identically, so there is no per-repository failure to report.
-func (c *client) searchMergedPRs(
-	ctx context.Context, repositories []models.Repository, mergedSince time.Time,
-) ([]PRResult, error) {
-	query := buildMergedPRsSearchQuery(repositories, mergedSince)
-
-	callCtx, cancel := context.WithTimeout(ctx, pullRequestListTimeout)
-	defer cancel()
-
-	var data aliasedData[searchResultsNode]
-	fieldErrors, err := c.graphql.Do(callCtx, query.text, query.variables, query.aliases, &data)
-	if err != nil {
-		return nil, mergedPRsFetchError(err)
-	}
-	if len(fieldErrors) > 0 {
-		return nil, mergedPRsFetchError(fieldErrors[0])
-	}
-	logRateLimit(data.RateLimit)
-
-	prResults := []PRResult{}
-	for _, alias := range query.aliases {
-		repository := query.repositoryByAlias[alias]
-		node := data.ByAlias[alias]
-		if node == nil {
-			return nil, mergedPRsFetchError(
-				fmt.Errorf("no data returned for %s", repository.GetPath()),
-			)
-		}
-		logTruncatedMergedPRs(repository, node.IssueCount)
-		prResults = append(
-			prResults, utilities.Map(node.Nodes, mergedPRResultMapper(repository))...,
-		)
-	}
-	return prResults, nil
-}
-
-func mergedPRsFetchError(err error) error {
-	return fmt.Errorf("error fetching merged pull requests: %w", err)
-}
-
-// The list is not paged, so a repository merging more than a page of PRs inside the window is
-// missing some of them, visible nowhere but the log.
-func logTruncatedMergedPRs(repository models.Repository, issueCount int) {
-	if issueCount <= mergedPRSearchPageSize {
-		return
-	}
-	log.Printf(
-		"%s merged %d pull requests within the window, listing only the newest %d of them",
-		repository.GetPath(), issueCount, mergedPRSearchPageSize,
-	)
-}
-
-func mergedPRResultMapper(repository models.Repository) func(node pullRequestNode) PRResult {
-	return func(node pullRequestNode) PRResult {
-		return PRResult{pr: mergedPullRequestFromNode(node), repository: repository}
-	}
-}
-
-// is:merged guarantees the state and the merged flag, which are therefore not selected.
-func mergedPullRequestFromNode(node pullRequestNode) *PullRequest {
-	pullRequest := pullRequestFromNode(node)
-	pullRequest.State = closedPullRequestState
-	pullRequest.Merged = true
-	return pullRequest
-}
-
-func assembleQuery(variableDeclarations, aliasSelections []string, fragment string) string {
-	return fmt.Sprintf(
-		"query(%s){\n  %s { cost remaining limit }\n%s\n}\n%s",
-		strings.Join(variableDeclarations, ","),
-		rateLimitField,
-		strings.Join(aliasSelections, "\n"),
-		fragment,
-	)
-}
-
-type rateLimitNode struct {
-	Cost      int `json:"cost"`
-	Remaining int `json:"remaining"`
-	Limit     int `json:"limit"`
-}
-
-type repositoryPullRequestsNode struct {
-	PullRequests connection[pullRequestNode] `json:"pullRequests"`
-}
-
-// The aliases are generated, so they are decoded by name rather than by struct field.
-type aliasedData[T any] struct {
-	RateLimit rateLimitNode
-	ByAlias   map[string]*T
-}
-
-func (d *aliasedData[T]) UnmarshalJSON(data []byte) error {
-	var fields map[string]json.RawMessage
-	if err := json.Unmarshal(data, &fields); err != nil {
-		return err
-	}
-
-	d.ByAlias = map[string]*T{}
-	for field, value := range fields {
-		if field == rateLimitField {
-			if err := json.Unmarshal(value, &d.RateLimit); err != nil {
-				return err
-			}
-			continue
-		}
-		var node *T
-		if err := json.Unmarshal(value, &node); err != nil {
-			return err
-		}
-		d.ByAlias[field] = node
-	}
-	return nil
-}
-
-func (c *client) listOpenPRs(
-	ctx context.Context, repositories []models.Repository,
-) ([]PRResult, error) {
-	query := buildListOpenPRsQuery(repositories)
-
-	callCtx, cancel := context.WithTimeout(ctx, pullRequestListTimeout)
-	defer cancel()
-
-	var data aliasedData[repositoryPullRequestsNode]
-	fieldErrors, err := c.graphql.Do(callCtx, query.text, query.variables, query.aliases, &data)
-	if err != nil {
-		return nil, listOpenPRsError(err, query.repositoryByAlias)
-	}
-	// A field error here loses every PR of one repository, which has always failed the whole list.
-	if len(fieldErrors) > 0 {
-		return nil, repositoryFetchError(query.repositoryByAlias[fieldErrors[0].alias], fieldErrors[0])
-	}
-	logRateLimit(data.RateLimit)
-
-	prResults := []PRResult{}
-	for _, alias := range query.aliases {
-		repository := query.repositoryByAlias[alias]
-		node := data.ByAlias[alias]
-		if node == nil {
-			return nil, repositoryFetchError(repository, errors.New("no data returned"))
-		}
-		prResults = append(
-			prResults,
-			utilities.Map(node.PullRequests.Nodes, openPRResultMapper(repository))...,
-		)
-	}
-	return prResults, nil
-}
-
-func listOpenPRsError(err error, repositoryByAlias map[string]models.Repository) error {
-	var repoError repositoryError
-	if errors.As(err, &repoError) {
-		if repository, isKnownAlias := repositoryByAlias[repoError.alias]; isKnownAlias {
-			return fmt.Errorf(
-				"repository %s/%s not found - check the repository name and permissions",
-				repository.Owner,
-				repository.Name,
-			)
-		}
-	}
-	return fmt.Errorf("error fetching pull requests: %w", err)
-}
-
-func repositoryFetchError(repository models.Repository, err error) error {
-	return fmt.Errorf(
-		"error fetching pull requests from %s/%s: %w", repository.Owner, repository.Name, err,
-	)
-}
-
-func logRateLimit(rateLimit rateLimitNode) {
-	log.Printf(
-		"GraphQL rate limit: cost %d, remaining %d of %d",
-		rateLimit.Cost, rateLimit.Remaining, rateLimit.Limit,
-	)
-}
-
-func openPRResultMapper(repository models.Repository) func(node pullRequestNode) PRResult {
-	return func(node pullRequestNode) PRResult {
-		return PRResult{pr: openPullRequestFromNode(node), repository: repository}
-	}
-}
-
-// states: OPEN guarantees the state, which is therefore not selected.
-func openPullRequestFromNode(node pullRequestNode) *PullRequest {
-	pullRequest := pullRequestFromNode(node)
-	pullRequest.State = openPullRequestState
-	pullRequest.Merged = false
-	return pullRequest
-}
-
-func pullRequestFromNode(node pullRequestNode) *PullRequest {
-	return &PullRequest{
-		Number:    node.Number,
-		Title:     node.Title,
-		HTMLURL:   node.URL,
-		CreatedAt: node.CreatedAt,
-		UpdatedAt: node.UpdatedAt,
-		State:     pullRequestStateFromNodeState(node.State),
-		Merged:    node.Merged,
-		Draft:     node.IsDraft,
-		Labels:    utilities.Map(node.Labels.Nodes, func(label labelNode) string { return label.Name }),
-		Author:    collaboratorFromAuthorNode(node.Author),
-		HeadSHA:   node.HeadRefOID,
-		MergedAt:  node.MergedAt,
-	}
-}
-
-// Only the two closed states close a PR, so an unexpected or missing state renders as open
-// rather than striking through every PR in the reminder.
-func pullRequestStateFromNodeState(nodeState string) string {
-	if nodeState == closedNodeState || nodeState == mergedNodeState {
-		return closedPullRequestState
-	}
-	return openPullRequestState
-}
-
-const enrichBatchSize = 25
-
-const approvedReviewState = "APPROVED"
-
-// A pending review is visible only to its own author, so it contributes no reviewer.
-const pendingReviewState = "PENDING"
-
-// The spread name and the definition are built together so that they cannot drift apart.
-type pullRequestFragment struct {
-	name string
-	text string
-}
-
-func newPullRequestFragment(name, selection string) pullRequestFragment {
-	return pullRequestFragment{
-		name: name,
-		text: fmt.Sprintf("fragment %s on PullRequest {\n%s\n}", name, selection),
-	}
-}
-
-// commits are selected for the PR tracker canvas, as the last-activity timestamp.
-const enrichedPullRequestSelection = `  number
-  commits(last: 1){ nodes { commit { oid committedDate } } }
-  reviews(first: 100){ nodes { state author { login __typename ... on User { name } } } }
-  comments(first: 100){ nodes { createdAt body author { login __typename ... on User { name } } } }`
-
-var enrichedPullRequestFragment = newPullRequestFragment("enrichedPr", enrichedPullRequestSelection)
-
-func buildPullRequestsQuery(
-	references []models.PullRequestRef, fragment pullRequestFragment,
-) graphqlQuery {
-	query := graphqlQuery{
-		variables:         map[string]any{},
-		aliases:           make([]string, len(references)),
-		repositoryByAlias: map[string]models.Repository{},
-		referenceByAlias:  map[string]models.PullRequestRef{},
-	}
-	variableDeclarations := make([]string, 0, 3*len(references))
-	pullRequestSelections := make([]string, len(references))
-
-	for index, reference := range references {
-		alias := fmt.Sprintf("p%d", index)
-		ownerVariable := fmt.Sprintf("owner%d", index)
-		nameVariable := fmt.Sprintf("name%d", index)
-		numberVariable := fmt.Sprintf("num%d", index)
-
-		query.aliases[index] = alias
-		query.repositoryByAlias[alias] = reference.Repository
-		query.referenceByAlias[alias] = reference
-		query.variables[ownerVariable] = reference.Repository.Owner
-		query.variables[nameVariable] = reference.Repository.Name
-		query.variables[numberVariable] = reference.Number
-		variableDeclarations = append(
-			variableDeclarations,
-			"$"+ownerVariable+":String!", "$"+nameVariable+":String!", "$"+numberVariable+":Int!",
-		)
-		pullRequestSelections[index] = fmt.Sprintf(
-			"  %s: repository(owner:$%s,name:$%s){ pullRequest(number:$%s){ ...%s } }",
-			alias, ownerVariable, nameVariable, numberVariable, fragment.name,
-		)
-	}
-
-	query.text = assembleQuery(variableDeclarations, pullRequestSelections, fragment.text)
-	return query
-}
-
-type pullRequestWrapperNode struct {
-	PullRequest *pullRequestNode `json:"pullRequest"`
-}
-
-// Fetches reviews and comments for the given PRs and returns them enriched, in input order.
-// Enrichment failures scoped to one PR are logged and leave that PR without reviewer info.
-func (c *client) enrichPRs(ctx context.Context, prResults []PRResult) ([]PR, error) {
-	log.Printf("\nFetching pull request reviews and comments for PRs")
-
-	batches := slices.Collect(slices.Chunk(prResults, enrichBatchSize))
-	enrichedBatches := make([][]PR, len(batches))
-
-	batchGroup, batchCtx := errgroup.WithContext(ctx)
-	batchGroup.SetLimit(defaultGitHubAPIConcurrencyLimit)
-
-	for index, batch := range batches {
-		batchGroup.Go(func() error {
-			enriched, err := c.enrichPRBatch(batchCtx, batch)
-			enrichedBatches[index] = enriched
-			return err
-		})
-	}
-	if err := batchGroup.Wait(); err != nil {
-		return nil, err
-	}
-	return utilities.FlatMap(enrichedBatches), nil
-}
-
-func (c *client) enrichPRBatch(ctx context.Context, batch []PRResult) ([]PR, error) {
-	query := buildPullRequestsQuery(
-		utilities.Map(batch, pullRequestRefOfResult), enrichedPullRequestFragment,
-	)
-
-	callCtx, cancel := context.WithTimeout(ctx, reviewsFetchTimeout)
-	defer cancel()
-
-	var data aliasedData[pullRequestWrapperNode]
-	fieldErrors, err := c.graphql.Do(callCtx, query.text, query.variables, query.aliases, &data)
-	if err != nil && !failsOnlyOnePullRequest(err) {
-		return nil, enrichPRsError(err, query.repositoryByAlias)
-	}
-	logRateLimit(data.RateLimit)
-
-	enriched := make([]PR, len(batch))
-	for index, result := range batch {
-		alias := query.aliases[index]
-		enriched[index] = enrichedPR(result, data.ByAlias[alias], errorsForAlias(alias, err, fieldErrors))
-	}
-	return enriched, nil
-}
-
-// Enrichment has never been able to fail the fetch, so a PR closed between the phases degrades
-// to the same per-PR path as a field error instead of aborting the run.
-func failsOnlyOnePullRequest(err error) bool {
-	var prError pullRequestError
-	return errors.As(err, &prError)
-}
-
-func errorsForAlias(alias string, hardError error, fieldErrors []fieldError) error {
-	aliasErrors := []error{}
-
-	var prError pullRequestError
-	if errors.As(hardError, &prError) && prError.alias == alias {
-		aliasErrors = append(aliasErrors, prError)
-	}
-	for _, err := range fieldErrors {
-		if err.alias == alias {
-			aliasErrors = append(aliasErrors, err)
-		}
-	}
-	return errors.Join(aliasErrors...)
-}
-
-func enrichPRsError(err error, repositoryByAlias map[string]models.Repository) error {
-	var repoError repositoryError
-	if errors.As(err, &repoError) {
-		if repository, isKnownAlias := repositoryByAlias[repoError.alias]; isKnownAlias {
-			return fmt.Errorf(
-				"error fetching reviews and comments from %s/%s: %w",
-				repository.Owner, repository.Name, err,
-			)
-		}
-	}
-	return fmt.Errorf("error fetching reviews and comments: %w", err)
-}
-
-func pullRequestRefOfResult(result PRResult) models.PullRequestRef {
-	return models.PullRequestRef{Repository: result.repository, Number: result.pr.GetNumber()}
-}
-
-var errNoPullRequestReturned = errors.New("no pull request returned")
-
-func enrichedPR(result PRResult, aliasNode *pullRequestWrapperNode, aliasError error) PR {
-	node, isEnriched := enrichedNode(aliasNode)
-	if !isEnriched {
-		aliasError = cmp.Or(aliasError, errNoPullRequestReturned)
-	}
-	logEnrichment(result.repository, result.pr.GetNumber(), node, aliasError)
-	return prWithReviewers(result.pr, result.repository, node)
-}
-
-// Reads the reviewer lists and the snooze off a PR's reviews and comments connections.
-func prWithReviewers(
-	pullRequest *PullRequest, repository models.Repository, node pullRequestNode,
-) PR {
-	submittedReviews := utilities.Filter(node.Reviews.Nodes, isSubmittedUserReview)
-	approvingReviews := utilities.Filter(submittedReviews, isApprovingReviewNode)
-	commentsFromUsers := utilities.Filter(node.Comments.Nodes, hasValidCommentAuthor)
-	timelineComments := utilities.Map(node.Comments.Nodes, timelineCommentFromNode)
-
-	approvedByUsers, commentedByUsers := deriveReviewers(
-		pullRequest.Author.Login,
-		utilities.Map(approvingReviews, reviewAuthor),
-		utilities.Map(submittedReviews, reviewAuthor),
-		utilities.Map(commentsFromUsers, commentAuthor),
-	)
-
-	return PR{
-		PullRequest:      pullRequestWithLastActivity(pullRequest, node),
-		Repository:       repository,
-		ApprovedByUsers:  approvedByUsers,
-		CommentedByUsers: commentedByUsers,
-		SnoozedUntil:     findActiveSnooze(timelineComments),
-	}
-}
-
-// Copies the PullRequest instead of writing the timestamp into the caller's one.
-func pullRequestWithLastActivity(pullRequest *PullRequest, node pullRequestNode) *PullRequest {
-	withLastActivity := *pullRequest
-	withLastActivity.LastActivityAt = lastActivityAt(pullRequest, node)
-	return &withLastActivity
-}
-
-// The update time overstates the last push, but bounds it from above, so a busy PR is never
-// mistaken for an inactive one. The head commit understates it instead: a PR opened today off an
-// old branch has an old head commit, so the commit date is raised to the PR's creation time.
-func lastActivityAt(pullRequest *PullRequest, node pullRequestNode) *time.Time {
-	if len(node.Commits.Nodes) > 0 {
-		committedDate := node.Commits.Nodes[0].Commit.CommittedDate
-		if createdAt := pullRequest.GetCreatedAt(); createdAt.After(committedDate) {
-			return &createdAt
-		}
-		return &committedDate
-	}
-	if updatedAt := pullRequest.GetUpdatedAt(); !updatedAt.IsZero() {
-		return &updatedAt
-	}
-	return nil
-}
-
-func enrichedNode(aliasNode *pullRequestWrapperNode) (pullRequestNode, bool) {
-	if aliasNode == nil || aliasNode.PullRequest == nil {
-		return pullRequestNode{}, false
-	}
-	return *aliasNode.PullRequest, true
-}
-
-func logEnrichment(repository models.Repository, number int, node pullRequestNode, err error) {
-	if err != nil {
-		log.Printf("Unable to fetch reviews/comments for PR #%d: %v", number, err)
-		return
-	}
-	log.Printf(
-		"Found %d reviews and %d timeline comments for PR %v/%d",
-		len(node.Reviews.Nodes), len(node.Comments.Nodes), repository, number,
-	)
-}
-
-func isSubmittedUserReview(review reviewNode) bool {
-	return review.State != pendingReviewState && hasValidAuthorNode(review.Author)
-}
-
-func isApprovingReviewNode(review reviewNode) bool {
-	return review.State == approvedReviewState
-}
-
-func reviewAuthor(review reviewNode) Collaborator {
-	return collaboratorFromAuthorNode(review.Author)
-}
-
-func hasValidCommentAuthor(comment commentNode) bool {
-	return hasValidAuthorNode(comment.Author)
-}
-
-func commentAuthor(comment commentNode) Collaborator {
-	return collaboratorFromAuthorNode(comment.Author)
-}
-
-// The referenced PRs may be closed or merged, so state and merged are selected alongside the
-// scalars, author and labels the open-PR listing provides in the "post" run mode.
-const fullPullRequestSelection = `  number title url isDraft createdAt updatedAt headRefOid state merged mergedAt
-  author { login __typename ... on User { name } }
-  labels(first: 100){ nodes { name } }
-  commits(last: 1){ nodes { commit { oid committedDate } } }
-  reviews(first: 100){ nodes { state author { login __typename ... on User { name } } } }
-  comments(first: 100){ nodes { createdAt body author { login __typename ... on User { name } } } }`
-
-var fullPullRequestFragment = newPullRequestFragment("fullPr", fullPullRequestSelection)
-
-// Fetches the referenced PRs with their reviews and comments, in input order. A PR that cannot
-// be fetched fails the call: unlike in FindOpenPRs there are no listed scalars to fall back to.
-func (c *client) getPRsByRef(ctx context.Context, references []models.PullRequestRef) ([]PR, error) {
-	batches := slices.Collect(slices.Chunk(references, enrichBatchSize))
-	fetchedBatches := make([][]PR, len(batches))
-
-	batchGroup, batchCtx := errgroup.WithContext(ctx)
-	batchGroup.SetLimit(defaultGitHubAPIConcurrencyLimit)
-
-	for index, batch := range batches {
-		batchGroup.Go(func() error {
-			prs, err := c.getPRBatchByRef(batchCtx, batch)
-			fetchedBatches[index] = prs
-			return err
-		})
-	}
-	if err := batchGroup.Wait(); err != nil {
-		return nil, err
-	}
-	return utilities.FlatMap(fetchedBatches), nil
-}
-
-func (c *client) getPRBatchByRef(
-	ctx context.Context, references []models.PullRequestRef,
-) ([]PR, error) {
-	query := buildPullRequestsQuery(references, fullPullRequestFragment)
-
-	callCtx, cancel := context.WithTimeout(ctx, reviewsFetchTimeout)
-	defer cancel()
-
-	var data aliasedData[pullRequestWrapperNode]
-	fieldErrors, err := c.graphql.Do(callCtx, query.text, query.variables, query.aliases, &data)
-	if err != nil {
-		return nil, getPRsError(err, query.referenceByAlias)
-	}
-	logRateLimit(data.RateLimit)
-
-	prs := make([]PR, len(references))
-	for index, reference := range references {
-		alias := query.aliases[index]
-		node, isFetched := enrichedNode(data.ByAlias[alias])
-		if !isFetched {
-			return nil, pullRequestFetchError(reference, errNoPullRequestReturned)
-		}
-		logEnrichment(
-			reference.Repository, reference.Number, node, errorsForAlias(alias, nil, fieldErrors),
-		)
-		prs[index] = prWithReviewers(pullRequestFromNode(node), reference.Repository, node)
-	}
-	return prs, nil
-}
-
-// Reports a missing PR separately from any other fetch failure.
-func getPRsError(err error, referenceByAlias map[string]models.PullRequestRef) error {
-	var prError pullRequestError
-	if !errors.As(err, &prError) {
-		return fmt.Errorf("error fetching pull requests: %w", err)
-	}
-	reference, isKnownAlias := referenceByAlias[prError.alias]
-	if !isKnownAlias {
-		return fmt.Errorf("error fetching pull requests: %w", err)
-	}
-	if prError.errorType == notFoundErrorType {
-		return pullRequestNotFoundError(reference)
-	}
-	return pullRequestFetchError(reference, err)
-}
-
-func pullRequestNotFoundError(reference models.PullRequestRef) error {
-	return fmt.Errorf(
-		"PR %s/%s/%d not found - check the path and permissions",
-		reference.Repository.Owner, reference.Repository.Name, reference.Number,
-	)
-}
-
-func pullRequestFetchError(reference models.PullRequestRef, err error) error {
-	return fmt.Errorf(
-		"error fetching pull request %s/%s/%d: %w",
-		reference.Repository.Owner, reference.Repository.Name, reference.Number, err,
-	)
-}
