+++ internal/apiclients/githubclient/graphqlmodels.go
@@ -1,11 +1,32 @@
 package githubclient
 
-import "time"
+import (
+	"time"
+
+	"github.com/hellej/pr-slack-reminder-action/internal/models"
+	"github.com/hellej/pr-slack-reminder-action/internal/utilities"
+)
 
 const (
 	botTypename  = "Bot"
 	userTypename = "User"
 )
+
+const (
+	openPullRequestState   = "open"
+	closedPullRequestState = "closed"
+)
+
+// GraphQL states of a PR node. CLOSED and MERGED are both "closed" to the pipeline.
+const (
+	closedNodeState = "CLOSED"
+	mergedNodeState = "MERGED"
+)
+
+const approvedReviewState = "APPROVED"
+
+// A pending review is visible only to its own author, so it contributes no reviewer.
+const pendingReviewState = "PENDING"
 
 // Nullable Actor; name is selected through "... on User { name }" so it is set for users only.
 type authorNode struct {
@@ -81,4 +102,106 @@ func timelineCommentFromNode(comment commentNode) TimelineComment {
 		Body:      comment.Body,
 		CreatedAt: comment.CreatedAt,
 	}
+}
+
+func pullRequestFromNode(node pullRequestNode) *PullRequest {
+	return &PullRequest{
+		Number:    node.Number,
+		Title:     node.Title,
+		HTMLURL:   node.URL,
+		CreatedAt: node.CreatedAt,
+		UpdatedAt: node.UpdatedAt,
+		State:     pullRequestStateFromNodeState(node.State),
+		Merged:    node.Merged,
+		Draft:     node.IsDraft,
+		Labels:    utilities.Map(node.Labels.Nodes, func(label labelNode) string { return label.Name }),
+		Author:    collaboratorFromAuthorNode(node.Author),
+		HeadSHA:   node.HeadRefOID,
+		MergedAt:  node.MergedAt,
+	}
+}
+
+// Only the two closed states close a PR, so an unexpected or missing state renders as open
+// rather than striking through every PR in the reminder.
+func pullRequestStateFromNodeState(nodeState string) string {
+	if nodeState == closedNodeState || nodeState == mergedNodeState {
+		return closedPullRequestState
+	}
+	return openPullRequestState
+}
+
+func enrichedNode(aliasNode *pullRequestWrapperNode) (pullRequestNode, bool) {
+	if aliasNode == nil || aliasNode.PullRequest == nil {
+		return pullRequestNode{}, false
+	}
+	return *aliasNode.PullRequest, true
+}
+
+// Reads the reviewer lists and the snooze off a PR's reviews and comments connections.
+func prWithReviewers(
+	pullRequest *PullRequest, repository models.Repository, node pullRequestNode,
+) PR {
+	submittedReviews := utilities.Filter(node.Reviews.Nodes, isSubmittedUserReview)
+	approvingReviews := utilities.Filter(submittedReviews, isApprovingReviewNode)
+	commentsFromUsers := utilities.Filter(node.Comments.Nodes, hasValidCommentAuthor)
+	timelineComments := utilities.Map(node.Comments.Nodes, timelineCommentFromNode)
+
+	approvedByUsers, commentedByUsers := deriveReviewers(
+		pullRequest.Author.Login,
+		utilities.Map(approvingReviews, reviewAuthor),
+		utilities.Map(submittedReviews, reviewAuthor),
+		utilities.Map(commentsFromUsers, commentAuthor),
+	)
+
+	return PR{
+		PullRequest:      pullRequestWithLastActivity(pullRequest, node),
+		Repository:       repository,
+		ApprovedByUsers:  approvedByUsers,
+		CommentedByUsers: commentedByUsers,
+		SnoozedUntil:     findActiveSnooze(timelineComments),
+	}
+}
+
+// Copies the PullRequest instead of writing the timestamp into the caller's one.
+func pullRequestWithLastActivity(pullRequest *PullRequest, node pullRequestNode) *PullRequest {
+	withLastActivity := *pullRequest
+	withLastActivity.LastActivityAt = lastActivityAt(pullRequest, node)
+	return &withLastActivity
+}
+
+// The update time overstates the last push, but bounds it from above, so a busy PR is never
+// mistaken for an inactive one. The head commit understates it instead: a PR opened today off an
+// old branch has an old head commit, so the commit date is raised to the PR's creation time.
+func lastActivityAt(pullRequest *PullRequest, node pullRequestNode) *time.Time {
+	if len(node.Commits.Nodes) > 0 {
+		committedDate := node.Commits.Nodes[0].Commit.CommittedDate
+		if createdAt := pullRequest.GetCreatedAt(); createdAt.After(committedDate) {
+			return &createdAt
+		}
+		return &committedDate
+	}
+	if updatedAt := pullRequest.GetUpdatedAt(); !updatedAt.IsZero() {
+		return &updatedAt
+	}
+	return nil
+}
+
+func isSubmittedUserReview(review reviewNode) bool {
+	return review.State != pendingReviewState && hasValidAuthorNode(review.Author)
+}
+
+func isApprovingReviewNode(review reviewNode) bool {
+	return review.State == approvedReviewState
+}
+
+func reviewAuthor(review reviewNode) Collaborator {
+	return collaboratorFromAuthorNode(review.Author)
+}
+
+func hasValidCommentAuthor(comment commentNode) bool {
+	return hasValidAuthorNode(comment.Author)
+}
+
+func commentAuthor(comment commentNode) Collaborator {
+	return collaboratorFromAuthorNode(comment.Author)
 }
