+++ internal/config/inputhelpers/inputhelpers.go
@@ -1,4 +1,4 @@
-// Package utilities provides helper functions for parsing GitHub Action inputs
+// Package inputhelpers provides helper functions for parsing GitHub Action inputs
 // from environment variables. It handles input name conversion, type parsing,
 // and structured data extraction from string inputs.
 package inputhelpers
